requirement.txt
blinker
boto3
botocore
click
Flask
importlib-metadata
itsdangerous
Jinja2
jmespath
MarkupSafe
python-dateutil
s3transfer
six
urllib3
Werkzeug
zipp
passlib
sqlalchemy
Flask-JWT-Extended
PyJWT
aniso8601
flask_restful
torch
transformers
Pillow
pyyaml
datasets